move schema to app
